moved the sign out to me tab
