+++ src/components/AuthContainer.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useEffect } from 'react';
-import { Button, SafeAreaView, Text } from 'react-native';
+import { Button, SafeAreaView } from 'react-native';
 import { withAuthenticator, AmplifyTheme } from 'aws-amplify-react-native';
 import { Auth, I18n } from 'aws-amplify';
 import { Translations } from '@aws-amplify/ui-components';
@@ -22,22 +22,12 @@ function App(): JSX.Element | null {
     }
   }, [appState, setAppState]);
 
-  const signOut = async () => {
-    try {
-      await Auth.signOut();
-      setAppState('Auth');
-    } catch (error) {
-      console.error('Error sign out: ', error);
-    }
-  };
-
   return (
     <>
       {appState === 'Onboarding' && <Onboarding />}
       {appState === 'App' && (
         <SafeAreaView style={{ flex: 1 }}>
           <NavFlow />
-          <Button title="Sign Out" onPress={signOut} />
         </SafeAreaView>
       )}
     </>
