+++ src/screens/ProfileScreen.tsx
@@ -1,10 +1,23 @@
-import React from 'react';
-import { Text, View } from 'react-native';
+import React, { useContext } from 'react';
+import { Button, View } from 'react-native';
+import { Auth } from 'aws-amplify';
+import { AppContext } from '../contexts/AppContext';
 
 export default function ProfileScreen() {
+  const { setAppState } = useContext(AppContext);
+
+  const signOut = async () => {
+    try {
+      await Auth.signOut();
+      setAppState('Auth');
+    } catch (error) {
+      console.error('Error signing out: ', error);
+    }
+  };
+
   return (
     <View>
-      <Text>Welcome to the profile screen!</Text>
+      <Button title="Sign Out" onPress={signOut} />
     </View>
   );
 }
